requirements.txt
llama-cpp-python>=0.3.0
grpcio>=1.68.0
grpcio-tools>=1.68.0
protobuf>=5.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.28.0
prometheus-client>=0.21.0
psutil>=6.0.0
pandas>=2.2.0
matplotlib>=3.9.0
seaborn>=0.13.0
pyyaml>=6.0.0
tqdm>=4.67.0
numpy>=2.0.0
huggingface_hub>=0.27.0
